added styling

+++ controllers/api/blogRoutes.js
@@ -1,20 +1,22 @@
 const express = require("express");
 const router = express.Router();
-const {User,Blog} = require("../models");
+const {User,Blog} = require("../../models");
+const withAuth = require('../../util/auth')
 
 
-//find all
+//find all blogs
 router.get("/", (req, res) => {
-  Blog.findAll({})
+  Blog.findAll({include:[User, Comment]})
     .then(dbBlogs => {
       res.json(dbBlogs);
     })
     .catch(err => {
       console.log(err);
-      res.status(500).json({ msg: "an error occured", err });
+      res.status(500).json({ msg: "an error occurred", err });
     });
 });
-//find one
+
+//find one specific blog
 router.get("/:id", (req, res) => {
   Blog.findByPk(req.params.id,{})
     .then(dbBlog => {
@@ -22,14 +24,14 @@ router.get("/:id", (req, res) => {
     })
     .catch(err => {
       console.log(err);
-      res.status(500).json({ msg: "an error occured", err });
+      res.status(500).json({ msg: "an error occurred", err });
     });
 });
 
-//create Blog
+//create new blog
 router.post("/", (req, res) => {
   if(!req.session.user){
-    return res.status(401).json({msg:"ya gotta login to create a blog post!"})
+    return res.status(401).json({msg:"Error - please log in"})
 }
   Blog.create({
     title:req.body.title,
@@ -41,27 +43,33 @@ router.post("/", (req, res) => {
     })
     .catch(err => {
       console.log(err);
-      res.status(500).json({ msg: "an error occured", err });
+      res.status(500).json({ msg: "an error occurred", err });
     });
 });
 
-//update Blog
+//update blog
 router.put("/:id", (req, res) => {
+  if(!req.session.user){
+    return res.status(401).json({msg:"Please login!"})
+  }
   Blog.update(req.body, {
-    where: {
-      id: req.params.id
-    }
-  }).then(updatedBlog => {
-    res.json(updatedBlog);
-  })
-  .catch(err => {
-    console.log(err);
-    res.status(500).json({ msg: "an error occured", err });
-  });
+      where: {
+        id: req.params.id
+      }
+    }).then(updatedBlog => {
+      res.json(updatedBlog);
+    })
+    .catch(err => {
+      console.log(err);
+      res.status(500).json({ msg: "an error occurred", err });
+    });
 });
 
-//delete a Blog
+//delete a blog
 router.delete("/:id", (req, res) => {
+  if(!req.session.user){
+    return res.status(401).json({msg:"Error - please log in"})
+  }
   Blog.destroy({
     where: {
       id: req.params.id
@@ -71,7 +79,7 @@ router.delete("/:id", (req, res) => {
   })
   .catch(err => {
     console.log(err);
-    res.status(500).json({ msg: "an error occured", err });
+    res.status(500).json({ msg: "an error occurred", err });
   });
 });
